% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/AddMeta.R
\name{AddMeta}
\alias{AddMeta}
\alias{.addMeta}
\alias{AddMeta,IRISFGM-method}
\title{AddMeta}
\usage{
AddMeta(object, ...)

.addMeta(object = NULL, meta.info = NULL)

\S4method{AddMeta}{IRISFGM}(object = NULL, meta.info = NULL)
}
\arguments{
\item{object}{input IRIS-FGM object}

\item{...}{other arguments passed to methods}

\item{meta.info}{meta information table should be a data frame with rows representing cell and coloumn representing different group condition}
}
\value{
It will add meta informatoin to IRISFGM.
}
\description{
This function can import cell annotation information to IRIS-FGM object.
}
\examples{
x <- matrix(rnorm(100),ncol = 10) 
colnames(x) <- paste0("cell",1:ncol(x))
rownames(x) <- paste0("gene",1:nrow(x))
my_meta <- data.frame(row.names = paste0("cell",1:ncol(x)),
cluster = c(rep("ClusterA",5),rep("ClusterB",5)))
object <- CreateIRISFGMObject(x)
object <- AddMeta(object, 
meta.info = my_meta)

}
