% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{BIC_LTMG}
\alias{BIC_LTMG}
\title{BIC_LTMG
BIC_LTMG}
\usage{
BIC_LTMG(y, rrr, Zcut)
}
\arguments{
\item{y}{input y}

\item{rrr}{input vector}

\item{Zcut}{input global z}
}
\value{
BIC_LTMG
}
\description{
BIC_LTMG
BIC_LTMG
}
