% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{BIC_ZIMG}
\alias{BIC_ZIMG}
\title{BIC_ZIMG fits different model
BIC_ZIMG fits different model}
\usage{
BIC_ZIMG(y, rrr, Zcut)
}
\arguments{
\item{y}{input vector}

\item{rrr}{input vector}

\item{Zcut}{global zcut}
}
\value{
BIC_ZIMG
}
\description{
BIC_ZIMG fits different model
BIC_ZIMG fits different model
}
