% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LTMG.R
\name{CalBinarySingleSignal}
\alias{CalBinarySingleSignal}
\alias{.CalBinarySingleSignal}
\alias{CalBinarySingleSignal,IRISFGM-method}
\title{CalBinarySingleSignal}
\usage{
CalBinarySingleSignal(object)

.CalBinarySingleSignal(object = NULL)

\S4method{CalBinarySingleSignal}{IRISFGM}(object = NULL)
}
\arguments{
\item{object}{Input IRIS-FGM object}
}
\value{
It will return a binary matrix based on LTMG signal matrix.
}
\description{
CalBinarySingleSignal

CalBinarySingleSignal
 
Binarizing single signal function via distinguishing zero or non-zero value based on LTMG matrix
}
\examples{
\dontrun{
object <- CalBinarySingleSignal(object)
}
}
