% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypePrediction.R
\name{CLUSTERING}
\alias{CLUSTERING}
\title{Packaging clustering method
This function is used for choosing clustering method}
\usage{
CLUSTERING(Raw, blocks, method = "MCL", K = NULL)
}
\arguments{
\item{Raw}{input raw discretized data which is chars file}

\item{blocks}{input block identified by IRISFGM}

\item{method}{chosse method, either MCL or SC.}

\item{K}{number of cluster}
}
\value{
clustering results
}
\description{
Packaging clustering method
This function is used for choosing clustering method
}
