% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{CreateIRISFGMObject}
\alias{CreateIRISFGMObject}
\title{CreateIRISFGMObject}
\usage{
CreateIRISFGMObject(
  x = input_matrix,
  min.cell = 0,
  min.gene = 0,
  LTMGr = new(Class = "LTMGr"),
  Bicluster = new(Class = "Bicluster")
)
}
\arguments{
\item{x}{Input expression matrix which should be a matrix or dataframe.}

\item{min.cell}{each gene should be expressed by at least this many cell.}

\item{min.gene}{each cell should express this many gene at least.}

\item{LTMGr}{Automatically create LTMG object.}

\item{Bicluster}{Automatically create Bicluster object.}
}
\value{
it should return a IRISFGM object of which structure can be found in tutorial.
}
\description{
Create IRIS-FGM object
}
\examples{
\dontrun{object <- CreateIRISCEMObject(x = input_matrix, min.cell = 0, min.gene =0}
}
