% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{DimReduce-class}
\alias{DimReduce-class}
\title{Create DimReduce object}
\description{
Create DimReduce object
}
\section{Slots}{

\describe{
\item{\code{PCA}}{ANY.}

\item{\code{UMAP}}{ANY.}

\item{\code{TSNE}}{ANY.}
}}

