% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{.CalBinaryMultiSignal}
\alias{.CalBinaryMultiSignal}
\title{calculate multisignal
This function is for calculating multisignal from LTMG signaling matrix.}
\usage{
.CalBinaryMultiSignal(object = NULL)
}
\arguments{
\item{object}{Input IRIS-FGM}
}
\value{
it will return a binary matrix
}
\description{
calculate multisignal
This function is for calculating multisignal from LTMG signaling matrix.
}
\examples{
\dontrun{object <- CalBinaryMultiSignal(object)}
}
