% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetwork.R
\name{.generateNetObject}
\alias{.generateNetObject}
\title{GenerateNetObject
Generate Net Object for the network use.}
\usage{
.generateNetObject(object, N.bicluster = c(1, 5), method = "spearman")
}
\arguments{
\item{object}{Input IRIS-FGM object.}

\item{N.bicluster}{Should be two integers indicating the number of two biclusters.}

\item{method}{Should be a statistical method to calculate edge weight based on expression data. It can be either "Spearman" (default) or "Pearson."}
}
\value{
get generateNetObject
}
\description{
GenerateNetObject
Generate Net Object for the network use.
}
