% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{.GetBinaryMultiSignal}
\alias{.GetBinaryMultiSignal}
\title{Get multisignal binary matrix
This function is for getting multisignal from LTMG signaling matrix.}
\usage{
.GetBinaryMultiSignal(object = NULL)
}
\arguments{
\item{object}{Input IRIS-FGM}
}
\value{
it will get a binary matrix
}
\description{
Get multisignal binary matrix
This function is for getting multisignal from LTMG signaling matrix.
}
\examples{
\dontrun{object <- CalBinaryMultiSignal(object)}
}
