% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bicluster.R
\name{.getBlock}
\alias{.getBlock}
\title{getblock function}
\usage{
.getBlock(object = NULL, keyword = "Conds")
}
\arguments{
\item{object}{IRISFGM object}

\item{keyword}{'Conds' for co-regulatory or 'Genes' for co-expression gene}
}
\value{
get blocks from biclustering results
}
\description{
get block from generated temporal files
}
