% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{.GetLTMGmatrix}
\alias{.GetLTMGmatrix}
\title{Get LTMG matrix}
\usage{
.GetLTMGmatrix(object)
}
\arguments{
\item{object}{Input IRIS-FGM object}
}
\description{
Get LTMG matrix
}
\examples{
\dontrun{GetLTMGmatrix(object)}
}
