% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichPathway.R
\name{.IDConvert}
\alias{.IDConvert}
\title{convert ID
convert ID}
\usage{
.IDConvert(genes.use = NULL, species = NULL)
}
\arguments{
\item{genes.use}{Provide gene list}

\item{species}{which species is in the provided gene list}
}
\value{
converted ID
}
\description{
convert ID
convert ID
}
