% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bicluster.R
\name{.runBiclusterBaseOnDiscretization}
\alias{.runBiclusterBaseOnDiscretization}
\title{Run Discretization
Generate temporal discretized file}
\usage{
.runBiclusterBaseOnDiscretization(
  object = NULL,
  OpenDual = TRUE,
  Extension = 1,
  NumBlockOutput = 100,
  BlockOverlap = 1,
  BlockCellMin = 15
)
}
\description{
Run Discretization
Generate temporal discretized file
}
