% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bicluster.R
\name{.runBiclusterBaseOnLTMG}
\alias{.runBiclusterBaseOnLTMG}
\title{RunBicusterBaseOnLTMG}
\usage{
.runBiclusterBaseOnLTMG(
  object = NULL,
  OpenDual = FALSE,
  Extension = 1,
  NumBlockOutput = 100,
  BlockOverlap = 0.9,
  BlockCellMin = 15
)
}
\description{
Run bicluster based on LTMG
}
