% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bicluster.R
\name{.runDiscretization}
\alias{.runDiscretization}
\title{run discretization}
\usage{
.runDiscretization(object = NULL, q = 0.06, LogTransformation = FALSE)
}
\arguments{
\item{object}{input IRIS-FGM object}

\item{q}{quantile number which is used as discretized cutoff. The bigger q means more cells will be categorized into 1 in terms of binarizing one gene.}
}
\value{
It will generate quantile based binary matrix.
}
\description{
run discretization based on Quantile method
}
