% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichPathway.R
\name{.RunGO}
\alias{.RunGO}
\title{RunGO
RunGO}
\usage{
.RunGO(genes.use = NULL, species = "mouse")
}
\arguments{
\item{genes.use}{Provide gene list}

\item{species}{which species is in the provided gene list}
}
\value{
GO pathway enrichment analysis
}
\description{
RunGO
RunGO
}
