% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichPathway.R
\name{.RunKEGG}
\alias{.RunKEGG}
\title{RunKEGG
RunKEGG}
\usage{
.RunKEGG(genes.use = NULL, species = "mouse")
}
\arguments{
\item{genes.use}{Provide gene list}

\item{species}{which species is in the provided gene list}
}
\value{
KEGG results
}
\description{
RunKEGG
RunKEGG
}
