% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetwork.R
\name{.separateBic}
\alias{.separateBic}
\title{separateBic
separate biclusters}
\usage{
.separateBic(object = NULL)
}
\arguments{
\item{object}{Input IRIS-FGM object.}
}
\value{
It will reture a list for shoring gene and cell.
}
\description{
separateBic
separate biclusters
}
