% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_object}
\alias{example_object}
\title{Example object.}
\format{
A data frame with 970 rows (gene) and 90 columns(cell):
\describe{
  \item{Raw_count}{slot for the original data}
  \item{Processed_count}{slot for the preprocessed data which was generated by normalization or imputation}
  ...
}
}
\source{
\url{https://bmbl.bmi.osumc.edu/downloadFiles/Yan_expression.txt}
}
\usage{
example_object
}
\description{
The example data was pre-generated IRISFGM object and it was generated from the partial data from Yan's data which contain 90 human embryo cells.
}
\keyword{datasets}
