% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DifferentialGene.R
\name{FindGlobalMarkers}
\alias{FindGlobalMarkers}
\alias{.findglobalMarkers}
\alias{FindGlobalMarkers,IRISFGM-method}
\title{This function is for finding global marker
FindGlobalMarkers is based on Seurat FindAllMarkers}
\usage{
FindGlobalMarkers(object, ...)

.findglobalMarkers(
  object = NULL,
  idents = "Seurat0.6",
  logfc.threshold = 0.25,
  test.use = "wilcox",
  only.pos = TRUE,
  random.seed = 1,
  min.pct = 0.1
)

\S4method{FindGlobalMarkers}{IRISFGM}(
  object = NULL,
  idents = "Seurat0.6",
  logfc.threshold = 0.25,
  test.use = "wilcox",
  only.pos = TRUE,
  random.seed = 1,
  min.pct = 0.1
)
}
\arguments{
\item{object}{input IRIS-FGM object}

\item{idents}{choose an idents for labelling cells}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25 Increasing logfc.threshold speeds up the function, but can miss weaker signals.}

\item{test.use}{same as \code{\link{FindAllMarkers}}}
}
\value{
Output is a differentially expressed gene list
}
\description{
This function is for finding global marker
FindGlobalMarkers is based on Seurat FindAllMarkers
}
\examples{
\dontrun{Global_marker <- FindGlobalMarkers(object)}
}
