% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{Fit_LTMG}
\alias{Fit_LTMG}
\title{Fit_LTMG
Fit_LTMG}
\usage{
Fit_LTMG(x, n, q, k, err = 1e-10)
}
\arguments{
\item{x}{input x}

\item{n}{input n}

\item{q}{input q}

\item{k}{input k}

\item{err}{random error}
}
\value{
Fit_LTMG
}
\description{
Fit_LTMG
Fit_LTMG
}
