% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LTMG.R
\name{GetBinaryMultiSignal}
\alias{GetBinaryMultiSignal}
\alias{.GetBinaryMultiSignal}
\alias{GetBinaryMultiSignal,IRISFGM-method}
\title{GetBinaryMultiSignal}
\usage{
GetBinaryMultiSignal(object, ...)

.GetBinaryMultiSignal(object = NULL)

\S4method{GetBinaryMultiSignal}{IRISFGM}(object = NULL)
}
\arguments{
\item{object}{Input IRIS-FGM}

\item{...}{other arguments passed to methods}
}
\value{
It will get a binary matrix based on LTMG signal matrix.
}
\description{
GetBinaryMultiSignal

GetBinaryMultiSignal
 
This function is for getting multisignal from LTMG signaling matrix.
}
\examples{
data(example_object)
Multisignal_matrix <- GetBinaryMultiSignal(example_object)
}
