% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LTMG.R
\name{GetBinarySingleSignal}
\alias{GetBinarySingleSignal}
\alias{.GetBinarySingleSignal}
\alias{GetBinarySingleSignal,IRISFGM-method}
\title{GetBinarySingleSignal
Get binary Single Signal matrix}
\usage{
GetBinarySingleSignal(object, ...)

.GetBinarySingleSignal(object = NULL)

\S4method{GetBinarySingleSignal}{IRISFGM}(object = NULL)
}
\arguments{
\item{object}{Input IRIS-FGM object}
}
\value{
It will export the Binarized matrix based on LTMG signal matrix.
}
\description{
GetBinarySingleSignal
Get binary Single Signal matrix
}
\examples{
\dontrun{
GetBinarySingleSignal(object)
}
}
