% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{IRISFGM-class}
\alias{IRISFGM-class}
\title{IRISFGM}
\description{
IRISFGM
}
\section{Slots}{

\describe{
\item{\code{Raw_count}}{ANY}

\item{\code{Processed_count}}{ANY}

\item{\code{MetaInfo}}{ANY}

\item{\code{Discretization}}{matrix.}

\item{\code{LTMG}}{LTMGr.}

\item{\code{BiCluster}}{Bicluster}
}}

