% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{IRISFGM}
\alias{IRISFGM}
\title{IRISFGM class}
\value{
IRIS-FGM object
}
\description{
IRISFGM class
}
\section{Slots}{

\describe{
\item{\code{MetaInfo}}{ANY.}

\item{\code{Discretization}}{matrix.}

\item{\code{Raw_count}}{ANY}

\item{\code{Processed_count}}{ANY}

\item{\code{BiCluster}}{Bicluster}

\item{\code{LTMG}}{LTMGr.}
}}

