% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{KS_ZIMG}
\alias{KS_ZIMG}
\title{KS_ZIMG
KS_ZIMG}
\usage{
KS_ZIMG(y, rrr, Zcut)
}
\arguments{
\item{y}{input y}

\item{rrr}{input rrr}

\item{Zcut}{input Zcut}
}
\value{
KS_ZIMG
}
\description{
KS_ZIMG
KS_ZIMG
}
