% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{LTMG}
\alias{LTMG}
\title{LTMG
LTMG}
\usage{
LTMG(VEC, Zcut_G, k = 5)
}
\arguments{
\item{VEC}{input vector}

\item{Zcut_G}{input Zcut}

\item{k}{input k as gene regulatory signal}
}
\description{
LTMG
LTMG
}
