% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypePrediction.R
\name{MCL}
\alias{MCL}
\title{MCL clustering
MCL clustering}
\usage{
MCL(Raw, blocks)
}
\arguments{
\item{Raw}{input data}

\item{blocks}{input blocks}
}
\value{
MCL clustering results
}
\description{
MCL clustering
MCL clustering
}
