% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMGSCA.R
\name{Pi_Zj_Zcut_new}
\alias{Pi_Zj_Zcut_new}
\title{Title
Title}
\usage{
Pi_Zj_Zcut_new(q, mean, sd, w10)
}
\arguments{
\item{q}{q}

\item{mean}{mean}

\item{sd}{sd}

\item{w10}{w10}
}
\description{
Title
Title
}
