% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DimensionReducntionBasedOnLTMG.R
\name{PlotDimension}
\alias{PlotDimension}
\alias{.plotDimension}
\alias{PlotDimension,IRISFGM-method}
\title{PlotDimension}
\usage{
PlotDimension(object, ...)

.plotDimension(object, reduction = "umap", pt_size = 1, idents = NULL)

\S4method{PlotDimension}{IRISFGM}(object, reduction = "umap", pt_size = 1, idents = NULL)
}
\arguments{
\item{object}{Input IRIS-FGM Object}

\item{...}{other arguments passed to methods}

\item{reduction}{Choose one of approaches for dimension reduction, including 'pca', 'tsne', 'umap'.}

\item{pt_size}{Point size, default is 0.}

\item{idents}{set current idents.}
}
\value{
generate plot on umap space.
}
\description{
Generate Umap and it requires user to input cell label index on console window.
}
\examples{
\dontrun{PlotDimension(object)}
}
