% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotHeatmap.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\alias{.plotHeatmap}
\alias{PlotHeatmap,IRISFGM-method}
\title{PlotHeatmap}
\usage{
PlotHeatmap(object, ...)

.plotHeatmap(
  object = object,
  N.bicluster = c(1, 5),
  show.overlap = FALSE,
  show.annotation = FALSE,
  show.clusters = FALSE
)

\S4method{PlotHeatmap}{IRISFGM}(
  object = object,
  N.bicluster = c(1, 5),
  show.overlap = FALSE,
  show.annotation = FALSE,
  show.clusters = FALSE
)
}
\arguments{
\item{object}{Input IRISFGM object}

\item{...}{other arguments passed to methods}

\item{N.bicluster}{Number of biclusters.}

\item{show.overlap}{Parameter (logic) indicates whether the figure shows the overlap part between two selected biclusters.}

\item{show.annotation}{Parameter (logic) indicates whether to show annotation (biclusters number and cell cluster labels).}

\item{show.clusters}{Parameter (logic) indicates whether to show the cell cluster label.}
}
\value{
It will generate a heatmap based on selected two FGMs.
}
\description{
PlotHeatmap

plot heatmap based on bicluster
}
\examples{
data(example_object)
PlotHeatmap(example_object,
N.bicluster = c(1,20),
show.annotation = TRUE, 
show.cluster =TRUE)

}
