% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialGene.R
\name{PlotMarkerHeatmap}
\alias{PlotMarkerHeatmap}
\title{PlotMarkerHeatmap will visualize global marker
This function will generate global marker gene heatmap}
\usage{
PlotMarkerHeatmap(
  Globalmarkers = NULL,
  object = NULL,
  idents = NULL,
  top.gene = 50,
  p.adj = 0.05,
  scale = "row",
  label.size = 1
)
}
\arguments{
\item{Globalmarkers}{output from FindGlobalMarkers}

\item{object}{input IRISFGM object}

\item{idents}{set current idents}

\item{top.gene}{this number of genes will be used for generating heatmap}

\item{p.adj}{adjusted pvalue cutoff for gene selection threshold}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "row" if symm false, and "none" otherwise.}

\item{label.size}{Number to decide label size}
}
\value{
heatmap
}
\description{
PlotMarkerHeatmap will visualize global marker
This function will generate global marker gene heatmap
}
\examples{
\dontrun{
markers <- FindGlobalMarkers(Global_marker,
object = object,
idents = "Seurat0.6",
top.gene = 50,
scale = "row")
}
}
