% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotNetwork.R
\name{PlotModuleNetwork}
\alias{PlotModuleNetwork}
\alias{.plotmodulenetwork}
\alias{PlotModuleNetwork,IRISFGM-method}
\title{PlotModuleNetwork}
\usage{
PlotModuleNetwork(object, ...)

.plotmodulenetwork(
  object = NULL,
  method = "spearman",
  N.bicluster = c(1, 5),
  cutoff.neg = -0.8,
  cutoff.pos = 0.95,
  layout = "circle",
  node.label = T,
  node.label.cex = 1
)

\S4method{PlotModuleNetwork}{IRISFGM}(
  object = NULL,
  method = "spearman",
  N.bicluster = c(1, 5),
  cutoff.neg = -0.8,
  cutoff.pos = 0.95,
  layout = "circle",
  node.label = T,
  node.label.cex = 1
)
}
\arguments{
\item{object}{Input object.}

\item{N.bicluster}{number of biclsuter to plot.}

\item{Node.color}{color of nodes. This parameter also accepts color codes, e.g. "#AE1503" or "darkred."}

\item{cutoff}{this parameter decide the cutoff of correlation}
}
\value{

}
\description{
This function will visualize co-expression gene network based on coorelation analysis. The nodes represent the gene module network from the selected bicluster. The size of the nodes indicates the degree of presence. The thickness of edges indicates the value of the correlation coefficient.
}
\examples{
\dontrun{object <- PlotModuleNetwork(object = NULL, N.bicluster = c(1,5), Node.color = "#E8E504", cutoff=0.7, node.label.cex = 1 )}
}
