% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotNetwork.R
\name{PlotNetwork}
\alias{PlotNetwork}
\alias{.plotnetwork}
\alias{PlotNetwork,IRISFGM-method}
\title{PlotNetwork}
\usage{
PlotNetwork(object, ...)

.plotnetwork(
  object,
  edge.by = "gene",
  lay.out = "linear",
  N.bicluster = c(1:20)
)

\S4method{PlotNetwork}{IRISFGM}(
  object,
  edge.by = "gene",
  lay.out = "linear",
  N.bicluster = c(1:20)
)
}
\arguments{
\item{object}{Input object}

\item{edge.by}{This parameter decide the edge by "cell" or by "gene". The default value is by gene.}

\item{lay.out}{The type of layout to create, include linear (default), circle, kk.}
}
\value{
It will generate a global network regarding overlapping genes or cells.
}
\description{
This function is for building the module- module network. Nodes mean individual module. Edges mean overlapped gene (or cell).
The weight of line show the number of gene (or cell).
}
\examples{
\dontrun{object <- PlotNetwork(object,edge.by = "gene",N.bicluster =c(1:20) )}
}
