% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bric.R
\name{qubic}
\alias{qubic}
\title{qubic}
\usage{
qubic(i, R = FALSE, F = FALSE, d = FALSE, f = 0.85, k = 13, c = 0.90, o = 5000)
}
\arguments{
\item{i}{input}

\item{N}{index}

\item{R}{index}

\item{Fa}{index}

\item{d}{index}

\item{D}{index}

\item{C}{index}

\item{n}{index}

\item{q}{index}

\item{f}{index}

\item{k}{index}

\item{c}{index}

\item{o}{index}
}
\description{
QUBIC Performs a QUalitative BIClustering.
}
