% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/InputData.R
\name{ReadFrom10X_h5}
\alias{ReadFrom10X_h5}
\alias{ReadFrom10X_h5,IRISFGM-method}
\title{Read 10X HDF5 file based on Seurat package}
\usage{
ReadFrom10X_h5(input = NULL, use.names = TRUE, unique.features = TRUE)

ReadFrom10X_h5(input = NULL, use.names = TRUE, unique.features = TRUE)

\S4method{ReadFrom10X_h5}{IRISFGM}(input = NULL, use.names = TRUE, unique.features = TRUE)
}
\arguments{
\item{input}{Input an HDF5 object}

\item{use.names}{Use barcode, default is true}

\item{unique.features}{use gene name, default is true}
}
\value{
The output from \code{\link{Read10X_h5}}

It will return a gene expression matrix.
}
\description{
This function provide a method for reading in HDF5 file from 10X platform.
}
\examples{
\dontrun{input_mat <- ReadFrom10X_h5(input= 'my.h5')}
}
