% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LTMG.R
\name{RunLTMG}
\alias{RunLTMG}
\alias{.RunLTMG}
\alias{RunLTMG,IRISFGM-method}
\title{RunLTMG}
\usage{
RunLTMG(object, ...)

.RunLTMG(object, Gene_use = NULL, k = 5)

\S4method{RunLTMG}{IRISFGM}(object, Gene_use = NULL, k = 5)
}
\arguments{
\item{object}{Input IRIS-FGM object}

\item{...}{other arguments passed to methods}

\item{Gene_use}{using X numebr of top variant gene. input a number, recommend 2000.}

\item{k}{Number of components.}
}
\value{
it will reture a LTMG signal matrix
}
\description{
We will use Left-truncated Mixture Gaussian distribution to model the regulatory signal of each gene. Parameter, 'Gene_use', decides number of top high variant gene for LTMG modeling, and here we use all genes.
}
\examples{
# If you want to explore DEG, we recommend you should use top 2000 highly variant gene. 
data("example_object")
example_object <- RunLTMG(example_object,
Gene_use = "all",
k = 5)

}
