% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypePrediction.R
\name{SC}
\alias{SC}
\title{spectral Clustering method 
spectral Clustering method}
\usage{
SC(Raw, blocks, K)
}
\arguments{
\item{Raw}{input}

\item{blocks}{input blocks}

\item{K}{number of clusters; this parameter is used for SC clustering method.}
}
\description{
spectral Clustering method 
spectral Clustering method
}
