% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMGSCA.R
\name{SeparateKRpkmNew}
\alias{SeparateKRpkmNew}
\title{Title
Title}
\usage{
SeparateKRpkmNew(x, n, q, k, err = 1e-10)
}
\arguments{
\item{x}{data, example: x<-runif(100,0,1)}

\item{n}{rounds}

\item{q}{cutoff}

\item{k}{k=1..5}

\item{err}{err}
}
\value{
a matrix contains pi, mean and sd
}
\description{
Title
Title
}
