% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMGSCA.R
\name{SeparateKRpkmNew2}
\alias{SeparateKRpkmNew2}
\title{SeparateKRpkmNew2
SeparateKRpkmNew2}
\usage{
SeparateKRpkmNew2(x, n, q, err = 1e-10)
}
\arguments{
\item{x}{x}

\item{n}{n}

\item{q}{q}

\item{err}{err}
}
\value{
SeparateKRpkmNew2
}
\description{
SeparateKRpkmNew2
SeparateKRpkmNew2
}
