% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{CreateIRISFGMObject}
\alias{CreateIRISFGMObject}
\title{CreateIRISFGMObject}
\arguments{
\item{x}{Input expression matrix which should be a matrix or dataframe.}

\item{min.cell}{each gene should be expressed by at least this many cell.}

\item{min.gene}{each cell should express this many gene at least.}

\item{LTMGr}{Automatically create LTMG object.}

\item{Bicluster}{Automatically create Bicluster object.}
}
\value{
it should return a IRISFGM object of which structure can be found in tutorial.
}
\description{
Create IRIS-FGM object
}
\details{
CreateIRISFGMObject
}
\examples{
x <- matrix(rnorm(100),ncol = 10) 
colnames(x) <- paste0("cell",1:ncol(x))
rownames(x) <- paste0("gene",1:nrow(x))
object <- CreateIRISFGMObject(x)
}
