% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/EnrichPathway.R
\name{DotPlotPathway}
\alias{DotPlotPathway}
\alias{.dotPlotPathway}
\alias{DotPlotPathway,IRISFGM-method}
\title{DotPlotPathway}
\usage{
DotPlotPathway(object, ...)

.dotPlotPathway(
  object = NULL,
  genes.source = c("CTS", "MC", "Bicluster"),
  showCategory = 20
)

\S4method{DotPlotPathway}{IRISFGM}(
  object = NULL,
  genes.source = c("CTS", "MC", "Bicluster"),
  showCategory = 20
)
}
\arguments{
\item{object}{Input IRIS-FGM object}

\item{...}{other arguments passed to methods}

\item{genes.source}{Decide the plot source either "CTS", 
"MC" or "Bicluster." "CTS" means DEGs from DEsingle label, 
"MC" means DEGs from MC label, and "Bicluster" means using gene module from the selected bicluster.}

\item{showCategory}{Show this number of pathway results.}
}
\value{
This function will generate dot plot for pathway enrichment results.
}
\description{
Plot dotplot for enrichment pathway
}
\examples{
data("example_object")
DotPlotPathway(example_object, genes.source = "module" )

}
