% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/EnrichPathway.R
\name{DotPlotPathway}
\alias{DotPlotPathway}
\alias{.dotplotpathway}
\alias{DotPlotPathway,IRISFGM-method}
\title{Plot dotplot for enrichment pathway}
\usage{
DotPlotPathway(object, ...)

.dotplotpathway(object = object, genes.source = c("CTS", "MC", "module"))

\S4method{DotPlotPathway}{IRISFGM}(object = object, genes.source = c("CTS", "MC", "module"))
}
\arguments{
\item{object}{Input IRIS-FGM object}

\item{genes.source}{decide the plot source either "CTS" or "Bicluster." "CTS" means from cell-type-specific DEGs, 
and "Bicluster means using gene module from the selected bicluster."}
}
\value{
dot plot
}
\description{
Plot dotplot for enrichment pathway
}
