% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DifferentialGene.R
\name{FindMarker}
\alias{FindMarker}
\alias{.findMarker}
\alias{FindMarker,IRISFGM-method}
\title{FindMarker}
\usage{
FindMarker(object, ...)

.findMarker(object, SimpleResult = T, FDR = 0.05)

\S4method{FindMarker}{IRISFGM}(object, SimpleResult = T, FDR = 0.05)
}
\arguments{
\item{object}{input IRISFGM object}

\item{...}{other arguments passed to methods}

\item{SimpleResult}{marker gene only output log fold change (LFC), p-value, and adjusted p-value.}

\item{FDR}{a number to specify the threshold of FDR, default by 0.05}
}
\description{
FindMarker

Find marker based on DEsingle method
Find marker based on DEsingle method
}
