% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LTMG.R
\name{GetBinarySingleSignal}
\alias{GetBinarySingleSignal}
\alias{.GetBinarySingleSignal}
\alias{GetBinarySingleSignal,IRISFGM-method}
\title{GetBinarySingleSignal}
\usage{
GetBinarySingleSignal(object, ...)

.GetBinarySingleSignal(object = NULL)

\S4method{GetBinarySingleSignal}{IRISFGM}(object = NULL)
}
\arguments{
\item{object}{Input IRIS-FGM object}

\item{...}{other arguments passed to methods}
}
\value{
It will export the Binarized matrix based on LTMG signal matrix.
}
\description{
GetBinarySingleSignal

GetBinarySingleSignal
Get binary Single Signal matrix
}
\examples{
Singlesignal_matrix <- GetBinarySingleSignal(example_object)

}
