% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{Global_Zcut}
\alias{Global_Zcut}
\title{Global_Zcut create Zcut 
Global_Zcut create Zcut}
\usage{
Global_Zcut(MAT, seed = NULL)
}
\arguments{
\item{MAT}{input matrix}

\item{seed}{set seed}
}
\description{
Global_Zcut create Zcut 
Global_Zcut create Zcut
}
