% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{IRISFGM}
\alias{IRISFGM}
\title{IRISFGM class}
\value{
IRIS-FGM object
}
\description{
IRISFGM class
}
\section{Slots}{

\describe{
\item{\code{raw_count}}{matrix.}

\item{\code{MetaInfo}}{ANY.}

\item{\code{processed_count}}{ANY.}

\item{\code{Discretization}}{matrix.}

\item{\code{LTMG}}{LTMGr.}

\item{\code{Bicluster}}{Bicluster.}
}}

