% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotHeatmap.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\alias{.plotHeatmap}
\alias{PlotHeatmap,IRISFGM-method}
\title{plot heatmap based on bicluster}
\usage{
PlotHeatmap(object, ...)

.plotHeatmap(
  object = object,
  N.bicluster = c(1, 5),
  show.overlap = F,
  seed = 123,
  show.annotation = F,
  show.clusters = F
)

\S4method{PlotHeatmap}{IRISFGM}(
  object = object,
  N.bicluster = c(1, 5),
  show.overlap = F,
  seed = 123,
  show.annotation = F,
  show.clusters = F
)
}
\arguments{
\item{object}{IRISFGM object}

\item{N.bicluster}{number of biclusters}

\item{show.overlap}{parameter indicates whether figure show the overlap part betwween two selected biclusters.}
}
\description{
plot heatmap based on bicluster
}
