% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotHeatmap.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\alias{.plotHeatmap}
\alias{PlotHeatmap,IRISFGM-method}
\title{plot heatmap based on bicluster}
\usage{
PlotHeatmap(object, ...)

.plotHeatmap(
  object = object,
  N.bicluster = c(1, 5),
  show.overlap = F,
  seed = 123,
  show.annotation = F,
  show.clusters = F
)

\S4method{PlotHeatmap}{IRISFGM}(
  object = object,
  N.bicluster = c(1, 5),
  show.overlap = F,
  seed = 123,
  show.annotation = F,
  show.clusters = F
)
}
\arguments{
\item{object}{Input IRISFGM object}

\item{N.bicluster}{Number of biclusters.}

\item{show.overlap}{Parameter (logic) indicates whether the figure shows the overlap part between two selected biclusters.}

\item{seed}{Set seeds for reproducibility.}

\item{show.annotation}{Parameter (logic) indicates whether to show annotation (biclusters number and cell cluster labels).}

\item{show.clusters}{Parameter (logic) indicates whether to show the cell cluster label.}
}
\description{
plot heatmap based on bicluster
}
