% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialGene.R
\name{PlotMarkerHeatmap}
\alias{PlotMarkerHeatmap}
\title{PlotMarkerHeatmap will visualize global marker
This function will generate global marker gene heatmap}
\usage{
PlotMarkerHeatmap(
  Globalmarkers = NULL,
  idents = "Seurat0.6",
  top.gene = 50,
  p.adj = 0.05,
  scale = "row"
)
}
\arguments{
\item{Globalmarkers}{output from FindGlobalMarkers}

\item{idents}{set current idents}

\item{top.gene}{this number of genes will be used for generating heatmap}

\item{p.adj}{adjusted pvalue cutoff for gene selection threshold}

\item{scale}{Scale by row or column.}
}
\value{
heatmap
}
\description{
PlotMarkerHeatmap will visualize global marker
This function will generate global marker gene heatmap
}
