% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotNetwork.R
\name{PlotModuleNetwork}
\alias{PlotModuleNetwork}
\alias{.plotmodulenetwork}
\alias{PlotModuleNetwork,IRISFGM-method}
\title{PlotModuleNetwork}
\usage{
PlotModuleNetwork(object, ...)

.plotmodulenetwork(
  object = NULL,
  method = "spearman",
  node.col = "orange",
  N.bicluster = c(1, 5),
  cutoff.neg = -0.8,
  cutoff.pos = 0.8,
  layout = "circle",
  node.label = TRUE,
  node.label.cex = 1
)

\S4method{PlotModuleNetwork}{IRISFGM}(
  object = NULL,
  method = "spearman",
  node.col = "orange",
  N.bicluster = c(1, 5),
  cutoff.neg = -0.8,
  cutoff.pos = 0.8,
  layout = "circle",
  node.label = TRUE,
  node.label.cex = 1
)
}
\arguments{
\item{object}{Input IRIS-FGM object.}

\item{method}{Should be a statistical method to calculate edge weight based on expression data. It can be either "Spearman" (default) or "Pearson."}

\item{node.col}{}

\item{N.bicluster}{Should be the two numbers of biclsuters.}

\item{cutoff.neg}{Should be a cutoff to show a negative correlation between two nodes (default: -0.8).}

\item{cutoff.pos}{Should be a cutoff to show a positive correlation between two nodes (default: 0.8).}

\item{layout}{Should be one type of layouts to show nodes' arrangement, including 'linear', 'star', 'circle'(default), 'gem', 'dh', 'graphopt', 'grid', 'mds', 
'randomly', 'fr', 'kk', 'drl', 'lgl'.}

\item{node.label}{Should be logic to show the nodes' label (default: TRUE).}

\item{node.label.cex}{Should be a number to control the label size.}
}
\value{

}
\description{
This function will visualize co-expression gene network based on selected two biclusters. 
The nodes represent the gene module network from the selected bicluster. 
The size of the nodes indicates the degree of presence.
 The thickness of edges indicates the value of the correlation coefficient.
}
\examples{
\dontrun{
object <- PlotModuleNetwork(object = NULL, 
N.bicluster = c(1,5), 
Node.color = '#E8E504', 
cutoff=0.7, 
node.label.cex = 1) 
}
}
