% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotNetwork.R
\name{PlotNetwork}
\alias{PlotNetwork}
\alias{PlotNetwork,IRISFGM-method}
\title{PlotNetwork}
\usage{
PlotNetwork(object, ...)

\S4method{PlotNetwork}{IRISFGM}(
  object,
  edge.by = "gene",
  lay.out = "linear",
  N.bicluster = c(1:20)
)
}
\arguments{
\item{object}{Input IRIS-FGM object.}

\item{...}{other arguments passed to methods}

\item{edge.by}{Should be "cell" or by "gene," indicating nodes label. The default value is by "gene."}

\item{lay.out}{Should be one type of layouts to show nodes' arrangement, including 'linear'(default), 'star', 'circle', 'gem', 'dh', 'graphopt', 'grid', 'mds', 
'randomly', 'fr', 'kk', 'drl', 'lgl'.}

\item{N.bicluster}{Should be two integers indicating the number of two biclusters.}
}
\value{
It will generate a global network regarding overlapping genes or cells.
}
\description{
PlotNetwork

PlotNetwork
This function is to plot the network for selected biclusters to show the genes (or cells) overlapping relations.
}
\examples{
\dontrun{object <- PlotNetwork(object,edge.by = 'gene',N.bicluster =c(1:20) )}
}
