% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PreprocessData.R
\name{ProcessData}
\alias{ProcessData}
\alias{.processData}
\alias{ProcessData,IRISFGM-method}
\title{ProcessData}
\usage{
ProcessData(object, ...)

.processData(
  object = NULL,
  normalization = "cpm",
  library.size = 1e+05,
  IsImputation = FALSE
)

\S4method{ProcessData}{IRISFGM}(
  object = NULL,
  normalization = "cpm",
  library.size = 1e+05,
  IsImputation = FALSE
)
}
\arguments{
\item{object}{Input IRISFGM object}

\item{...}{other arguments passed to methods}

\item{normalization}{two options: (1)library size noralization by using library size factor: 1e6, equal to CPM (count-per-million-reads), or 
(2) using 'scran' normalization method.}

\item{library.size}{Sets the scale factor for cpm normalization}

\item{IsImputation}{imputation method is provided by DrImpute. Default is FALSE.}
}
\value{
It will processdata by normalization and imputation.
}
\description{
Process data via normalization and imputation.
}
\examples{
 
data("example_object")
example_object <- ProcessData(example_object,
normalization = "cpm")
}
