% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bric.R
\name{qubic}
\alias{qubic}
\title{qubic}
\usage{
qubic(
  i = NULL,
  N = FALSE,
  R = FALSE,
  Fa = FALSE,
  d = FALSE,
  D = FALSE,
  C = FALSE,
  n = FALSE,
  q = 0.05,
  f = 0.85,
  k = 13,
  c = 0.9,
  o = 100
)
}
\arguments{
\item{i}{input}

\item{N}{index}

\item{R}{index}

\item{Fa}{index}

\item{d}{index}

\item{D}{index}

\item{C}{index}

\item{n}{index}

\item{q}{index}

\item{f}{index}

\item{k}{index}

\item{c}{index}

\item{o}{index}
}
\value{
qubic results
}
\description{
QUBIC Performs a QUalitative BIClustering.
}
