% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/InputData.R
\name{ReadFrom10X_folder}
\alias{ReadFrom10X_folder}
\alias{ReadFrom10X_folder,IRISFGM-method}
\title{Read 10X folder based on Seurat package}
\usage{
ReadFrom10X_folder(input.dir = NULL)

ReadFrom10X_folder(input.dir = NULL)

\S4method{ReadFrom10X_folder}{IRISFGM}(input.dir = NULL)
}
\arguments{
\item{input.dir}{Input directory. It should contain three file, including barcode file, feature file, and sparse counts matrix.}
}
\value{
The output from \code{\link{Read10X}}
}
\description{
This function provide a method for reading in a folder from 10X platform. In this folder, it should contain three files: barcode, matrix, and gene.
}
\examples{
\dontrun{input_mat <- ReadFrom10X_folder(input.dir = 'my_path_to_folder')}
}
