% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Bicluster.R
\name{RunBicluster}
\alias{RunBicluster}
\alias{.runBicluster}
\alias{RunBicluster,IRISFGM-method}
\title{RunBicluster}
\usage{
RunBicluster(object, ...)

.runBicluster(
  object = NULL,
  DiscretizationModel = "Quantile",
  OpenDual = FALSE,
  Extension = 1,
  NumBlockOutput = 100,
  BlockOverlap = 0.7,
  BlockCellMin = 15
)

\S4method{RunBicluster}{IRISFGM}(
  object = NULL,
  DiscretizationModel = "Quantile",
  OpenDual = FALSE,
  Extension = 1,
  NumBlockOutput = 100,
  BlockOverlap = 0.7,
  BlockCellMin = 15
)
}
\arguments{
\item{object}{input IRIS-FGM object}

\item{...}{other arguments passed to methods}

\item{DiscretizationModel}{use different discretization method, including 'Quantile' and 'LTMG.'}

\item{OpenDual}{the flag using the lower bound of condition number. Default: 5 percent of the gene number in current bicluster.}

\item{Extension}{consistency level of the block (0.5-1.0], the minimum ratio between the number of identical valid symbols in a column and the total number of rows in the output. Default: 1.0.}

\item{NumBlockOutput}{number of blocks to report. Default: 100.}

\item{BlockOverlap}{filtering overlapping blocks. Default: 0.7.}

\item{BlockCellMin}{minimum column width of the block. Default: 15 columns.}
}
\value{
It will generate a temporal file on local directory for processed data named 'tmp_expression.txt', discretized file named 
'tmp_expression.txt.chars', and biclsuter block named 'tmp_expression.txt.chars.block'.
}
\description{
This function will identify the Biclusters based on LTMG or Quantile normalization
}
\examples{
# based on LTMG discretization
# data(example_data)
# example_object <- RunBicluster(example_object, 
# DiscretizationModel = 'LTMG',
# OpenDual = F,
# NumBlockOutput = 1000, 
# BlockOverlap = 0.7, 
# BlockCellMin = 15)

# based on quantile discretization
# data(example_data)
# example_object <- RunBicluster(example_object, 
# DiscretizationModel = 'Quantile',
# OpenDual = F, 
# NumBlockOutput = 1000, 
# BlockOverlap = 0.7, 
# BlockCellMin = 15)

}
