% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionReducntionBasedOnLTMG.R
\name{RundimensionReduction}
\alias{RundimensionReduction}
\alias{.runDimensionReduction}
\title{Run dimension reduction based on seurat method}
\usage{
.runDimensionReduction(
  object,
  mat.source = c("LTMG", "UMImatrix"),
  reduction = "umap",
  dims = 1:15,
  perplexity = 15,
  seed = 1
)
}
\arguments{
\item{object}{input IRIS-FGM}

\item{mat.source}{choose source data for running this function either from LTMG signal matrix or from processed data. Values of this parameter are "LTMG" and "UMImatrix"}

\item{reduction}{select a method for dimension reduction, including umap, tsne, and pca.}

\item{dims}{select the number of PCs from PCA results to perform the following dimension reduction and cell clustering.}
}
\value{
This function will generate pca, tsne, or umap dimension reduction results.
}
\description{
This function is based on the Seurat package to perform dimension reduction. The input matirx is the LTMG signaling matrix.
}
\examples{
\dontrun{obejct <- RunDimensionReduction(object,mat.source= "LTMG", reduction = "umap", dims = 1:15 ,perplexity = 15, seed = 1)}
}
