% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LTMG.R
\name{RunLTMG}
\alias{RunLTMG}
\alias{.RunLTMG}
\alias{RunLTMG,IRISFGM-method}
\title{Run LTMG module}
\usage{
RunLTMG(object, ...)

.RunLTMG(object, Gene_use = NULL, seed = 123, k = 5)

\S4method{RunLTMG}{IRISFGM}(object, Gene_use = NULL, seed = 123, k = 5)
}
\arguments{
\item{object}{Input IRIS-FGM object}

\item{Gene_use}{using X numebr of top variant gene. input a number, recommend 2000.}

\item{seed}{Set seeds for reproducibility}

\item{k}{Number of components.}
}
\value{
it will reture a LTMG signal matrix
}
\description{
We will use Left-truncated Mixture Gaussian distribution to model the regulatory signal of each gene. Parameter, 'Gene_use', decides number of top high variant gene for LTMG modeling, and here we use all genes.
}
\examples{
# If you want to explore DEG, we recommend you should use top 2000 highly variant gene. 
\dontrun{
object <- RunLTMG(object,
Gene_use = 2000, 
seed = 123, 
k = 5)
}
# If you want to run bicluster based on LTMG model, we recommend you should use all genes.
\dontrun{
object <- RunLTMG(object,
Gene_use ='all', 
seed = 123, 
k = 5)}
}
