% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/AddMeta.R
\name{SubsetData}
\alias{SubsetData}
\alias{.subset_data}
\alias{SubsetData,IRISFGM-method}
\title{Subset data by number of count and numebr of gene.}
\usage{
SubsetData(object, ...)

.subset_data(
  object,
  nFeature.upper = Inf,
  nFeature.lower = -Inf,
  Counts.upper = Inf,
  Counts.lower = -Inf
)

\S4method{SubsetData}{IRISFGM}(
  object,
  nFeature.upper = Inf,
  nFeature.lower = -Inf,
  Counts.upper = Inf,
  Counts.lower = -Inf
)
}
\arguments{
\item{object}{input IRIS-FGM object}

\item{nFeature.upper}{select upper limit for number of feature}

\item{nFeature.lower}{select lower limit for number of feature}

\item{Counts.upper}{select upper limit for number of UMI counts}

\item{Counts.lower}{select lower limit for number of UMI counts}
}
\value{
it will filter out some cell regarding threshhold.
}
\description{
This function is used for filtering out low-quality data based on previous result generated from \code{\link{PlotMeta}}
}
\examples{
# Use Yan's data which posts on github tutorial
\dontrun{ 
object <- SubsetData(object, 
    nFeature.upper=15000,
    nFeature.lower=8000, 
    Counts.upper=700000,
    Counts.lower=400000)}

}
